Have begun moving draw responsibilities out of WorldRenderer
All Entities now extend an abstract draw method that contains the logic
to draw itself.  This may or may not be the "normal" way to do this, but
it makes sense to me for drawn objects to contain their own logic to
draw themselves.

This also reduces the load on the bloated WorldRenderer.render() method.