Ant's now "clickable"
If user clicks on an Ant, the screen will zoom to Ant and follow it.
User can then click anywhere to unselect.

Future feature to add: ant thoughts/food levels.  Also consider making a
seperate window pop up that follows the ant so as to not mess with main
UI.