Update to Menu creation
Button size now tied to the size of the actual device screen size.  This
will probably be changed again in the future, but for now it works
pretty well transitioning between desktop and a standard android device.
I imagine the buttons will seem a bit big on a tablet, however.