Now using TexturePacker/Atlas
Before, the textures wouldn't load on older Android systems because they
weren't a power of 2 (requirment for OpenGL 1.0).  Now they work.

I've also moved the loading of textures outside of the indvidual classes
- it was stupid to be loading them dozens of time in the first place.
The problem now is that the render loop of WorldRenderer is much more
complicated and hard to read.  Need to find a better way to abstract it
all out.