Moved Texture allocation to Level
It makes more sense to me to load all of our resources (both entity
objects and their respective textures) specific to each level.

The problem is now I need to figure out a way to add new ants in the
Anthill.class createAnt() because Ant's require a texture.

The reason I took the textures out of each class at all was because I'm
now using a TextureAtlas and it seems silly to create one for each
concrete class just to get access to their respective texture.  And also
because, as I mentioned above, it feels more 'right' to instantiate that
value outside of the class.

I'll have to think more about this and look into it in the future.  I'm
worried about creating multiple textures on accident and thought I was
creating a memory leak because of it according to the logcat information
about "Freeing memory".  It turns out that is shown no matter what app
is at the forefront - it may just be a normal thing and never dipped
below 40% (30% is when it's supposed to create stuttering) at 200 ant's
created.  So...more research is needed.

Until then, I think the code looks much cleaner this way.